update blowout/dispense tests, factor out dispenseUpdateLiquidState test
